# [lldb] Optimize target read from LLDBMemoryReader::readString