@kastiglione 2021-01-08
Jason was surprised too. My take away is that it's not just allocation, but whatever (fixed or variable) construction costs vector has as well. Vector also has costs for destructing too.