@adrian-prantl 2021-01-08
It really emphasizes just how important profiling is — we can't trust our intuition when it comes to performance.